Move theme commands and tests to separate files.
