+++ tests/phpunit.xml
@@ -24,6 +24,7 @@
       <file>backdrop/RolesCommandsTest.php</file>
       <file>backdrop/StateCommandsTest.php</file>
       <!-- There's no Status command test, since that command is used in other tests already. -->
+      <file>backdrop/ThemeCommandsTest.php</file>
       <file>backdrop/UpdateCommandsTest.php</file>
       <file>backdrop/UserCommandsTest.php</file>
     </testsuite>
